+++ Changelog.md
@@ -2,6 +2,16 @@
 # CyberFilesRewrite Changelog
 I'll try my best to keep track of all changes to CyberFilesRewrite, big and small, right here in this changelog. I do most of my web development late at night, so the dates here tend to extend into the following day.
 
+## 1.6.0 - September 22nd, 2021
+* Reduced the width of the Modified column from 160px to 140px
+* Reduced the max width of the file list from 1200px to 1100px
+* Added the Type column in the file list
+    * This is only visible on screen widths larger than 950px
+* Added access and error logging
+    * Logs are stored in `/_cyberfiles/private/logs`, and are separated into files for each day
+* Added sorting by file type (and accompanying config options)
+    * This option actually sorts by file extension, as opposed to the custom language-defined type
+
 ## 1.5.0 - September 21st, 2021
 * Fixed the Home icon in recents
 * Added the `videoProgressSave.prompt` config option
@@ -11,7 +21,7 @@ I'll try my best to keep track of all changes to CyberFilesRewrite, big and smal
         * Clicking the header of the current sort column will reverse the direction of the sort
     * The chosen sort order is stored per-directory, and remains until changed
 * Increased the max width of the file list from 1000px to 1200px
-    * This is in preporation for a file type column that'll only appear at larger widths
+    * This is in preparation for a file type column that'll only appear at larger widths
 
 ## 1.4.0 - September 20th, 2021
 * Added more padding to the bottom of h1-h6 elements to go along with yesterday's line height changes